requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2
django-mongoengine==0.5.4
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
djongo==1.3.6
dnspython==2.2.1
google==3.0.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.20.4
idna==3.3
mongo==0.2.0
mongoengine==0.24.1
oauthlib==3.2.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pymongo==3.12.1
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.1.0
social-auth-storage-mongoengine==1.0.1
soupsieve==2.3.2.post1
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.9
